requirements.txt
absl-py==0.8.1
astor==0.8.0
cachetools==3.1.1
certifi==2019.11.28
chardet==3.0.4
Click==7.0
cycler==0.10.0
Flask==1.1.1
gast==0.2.2
google-auth==1.7.2
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
grpcio==1.25.0
gunicorn==20.0.4
h5py==2.10.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.1
numpy==1.17.4
oauthlib==3.1.0
opt-einsum==3.1.0
Pillow==6.2.1
protobuf==3.11.1
pyasn1==0.4.8
pyasn1-modules==0.2.7
pyparsing==2.4.5
python-dateutil==2.8.1
PyYAML==5.2
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
scipy==1.3.3
six==1.13.0
tensorboard==2.0.2
tensorflow==2.0.0
tensorflow-estimator==2.0.1
termcolor==1.1.0
urllib3==1.25.7
Werkzeug==0.16.0
wrapt==1.11.2
